+++ apps/meteor/client/navbar/NavBarPagesGroup/actions/CreateTeamModal.spec.tsx
@@ -4,9 +4,7 @@ import userEvent from '@testing-library/user-event';
 
 import CreateTeamModal from './CreateTeamModal';
 
-jest.mock('../../../lib/utils/goToRoomById', () => ({
-	goToRoomById: jest.fn(),
-}));
+jest.mock('../../../lib/rooms/roomCoordinator', () => ({}));
 
 describe('CreateTeamModal', () => {
 	it('should render with encryption option disabled and set to off when E2E_Enable=false and E2E_Enabled_Default_PrivateRooms=false', async () => {