+++ apps/meteor/client/views/room/contextualBar/RoomFiles/components/FileItem.tsx
@@ -4,6 +4,7 @@ import { Box } from '@rocket.chat/fuselage';
 import FileItemIcon from './FileItemIcon';
 import FileItemMenu from './FileItemMenu';
 import ImageItem from './ImageItem';
+import { normalizeUsername } from '../../../../../../lib/utils/normalizeUsername';
 import { useDownloadFromServiceWorker } from '../../../../../hooks/useDownloadFromServiceWorker';
 import { useFormatDateAndTime } from '../../../../../hooks/useFormatDateAndTime';
 
@@ -17,11 +18,12 @@ const FileItem = ({ fileData, onClickDelete }: FileItemProps) => {
 	const { _id, path, name, uploadedAt, type, typeGroup, user } = fileData;
 
 	const encryptedAnchorProps = useDownloadFromServiceWorker(path || '', name);
+	const normalizedUsername = user?.username ? normalizeUsername(user.username) : undefined;
 
 	return (
 		<>
 			{typeGroup === 'image' ? (
-				<ImageItem id={_id} url={path} name={name} username={user?.username} timestamp={format(uploadedAt)} />
+				<ImageItem id={_id} url={path} name={name} username={normalizedUsername} timestamp={format(uploadedAt)} />
 			) : (
 				<Box
 					is='a'
@@ -46,7 +48,7 @@ const FileItem = ({ fileData, onClickDelete }: FileItemProps) => {
 						</Box>
 						{user?.username && (
 							<Box withTruncatedText color='hint' fontScale='p2'>
-								@{user?.username}
+								@{normalizedUsername}
 							</Box>
 						)}
 						<Box color='hint' fontScale='micro'>