+++ ee/packages/media-calls/src/server/CallDirector.ts
@@ -11,6 +11,7 @@ import type { InsertionModel } from '@rocket.chat/model-typings';
 import { MediaCallNegotiations, MediaCalls } from '@rocket.chat/models';
 
 import { getCastDirector, getMediaCallServer } from './injection';
+import { DEFAULT_CALL_FEATURES } from '../constants';
 import type { IMediaCallAgent } from '../definition/IMediaCallAgent';
 import type { IMediaCallCastDirector } from '../definition/IMediaCallCastDirector';
 import type { InternalCallParams, MediaCallHeader } from '../definition/common';
@@ -79,7 +80,7 @@ class MediaCallDirector {
 		this.scheduleExpirationCheckByCallId(call._id);
 
 		const updatedCall = await MediaCalls.findOneById(call._id, { projection: { features: 1 } });
-		const features = (updatedCall?.features || ['audio']) as CallFeature[];
+		const features = (updatedCall?.features || DEFAULT_CALL_FEATURES) as CallFeature[];
 
 		await calleeAgent.onCallAccepted(call._id, { signedContractId: data.calleeContractId, features });
 		await calleeAgent.oppositeAgent?.onCallAccepted(call._id, { signedContractId: call.caller.contractId, features });
@@ -177,17 +178,7 @@ class MediaCallDirector {
 	}
 
 	public async createCall(params: CreateCallParams): Promise<IMediaCall> {
-		const {
-			caller,
-			callee,
-			requestedCallId,
-			requestedService,
-			callerAgent,
-			calleeAgent,
-			parentCallId,
-			requestedBy,
-			features = ['audio'],
-		} = params;
+		const { caller, callee, requestedCallId, requestedService, callerAgent, calleeAgent, parentCallId, requestedBy, features } = params;
 
 		// The caller must always have a contract to create the call
 		if (!caller.contractId) {
@@ -212,6 +203,8 @@ class MediaCallDirector {
 		callerAgent.oppositeAgent = calleeAgent;
 		calleeAgent.oppositeAgent = callerAgent;
 
+		const allowedFeatures = features.filter((feature) => getMediaCallServer().isFeatureAvailableForUser(caller.id, feature));
+
 		const call: Omit<IMediaCall, '_id' | '_updatedAt'> = {
 			service,
 			kind: 'direct',
@@ -234,7 +227,7 @@ class MediaCallDirector {
 			...(requestedCallId && { callerRequestedId: requestedCallId }),
 			...(parentCallId && { parentCallId }),
 
-			features,
+			features: allowedFeatures,
 		};
 
 		logger.debug({ msg: 'creating call', call });